Descriptors: changed devID from 32-bits to 64-bits (#171)
* In serialized form of nixlBasicDesc, it was using 24 bytes anyways on 64-bit
  systems, so no memory savings, while going to u64 can be useful for some backends.
* Also used constexpr as we moved to c++17, and now empty copyMeta can be removed
  from nixlBasicDesc.

+++ src/bindings/python/nixl_bindings.cpp
@@ -163,7 +163,7 @@ PYBIND11_MODULE(_bindings, m) {
         .def(py::init([](nixl_mem_t mem, std::vector<py::tuple> descs, bool sorted) {
                 nixl_xfer_dlist_t new_list(mem, sorted, descs.size());
                 for(long unsigned int i = 0; i<descs.size(); i++)
-                    new_list[i] = nixlBasicDesc(descs[i][0].cast<uintptr_t>(), descs[i][1].cast<size_t>(), descs[i][2].cast<uint32_t>());
+                    new_list[i] = nixlBasicDesc(descs[i][0].cast<uintptr_t>(), descs[i][1].cast<size_t>(), descs[i][2].cast<uint64_t>());
                 if (sorted) new_list.verifySorted();
                 return new_list;
             }), py::arg("type"), py::arg("descs"), py::arg("sorted")=false)
@@ -173,8 +173,8 @@ PYBIND11_MODULE(_bindings, m) {
         .def("isSorted", &nixl_xfer_dlist_t::isSorted)
         .def(py::self == py::self)
         .def("__getitem__", [](nixl_xfer_dlist_t &list, unsigned int i) ->
-              std::tuple<uintptr_t, size_t, uint32_t> {
-                    std::tuple<uintptr_t, size_t, uint32_t> ret;
+              std::tuple<uintptr_t, size_t, uint64_t> {
+                    std::tuple<uintptr_t, size_t, uint64_t> ret;
                     nixlBasicDesc desc = list[i];
                     std::get<0>(ret) = desc.addr;
                     std::get<1>(ret) = desc.len;
@@ -182,17 +182,17 @@ PYBIND11_MODULE(_bindings, m) {
                     return ret;
               })
         .def("__setitem__", [](nixl_xfer_dlist_t &list, unsigned int i, const py::tuple &desc) {
-                list[i] = nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint32_t>());
+                list[i] = nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint64_t>());
             })
         .def("addDesc", [](nixl_xfer_dlist_t &list, const py::tuple &desc) {
-                list.addDesc(nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint32_t>()));
+                list.addDesc(nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint64_t>()));
             })
         .def("append", [](nixl_xfer_dlist_t &list, const py::tuple &desc) {
-                list.addDesc(nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint32_t>()));
+                list.addDesc(nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint64_t>()));
             })
         .def("index", [](nixl_xfer_dlist_t &list, const py::tuple &desc) {
                 int ret = (nixl_status_t) list.getIndex(nixlBasicDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(),
-                                                  desc[2].cast<uint32_t>()));
+                                                  desc[2].cast<uint64_t>()));
                 if(ret < 0) throw_nixl_exception((nixl_status_t) ret);
                 return (int) ret;
             })
@@ -220,7 +220,7 @@ PYBIND11_MODULE(_bindings, m) {
         .def(py::init([](nixl_mem_t mem, std::vector<py::tuple> descs, bool sorted) {
                 nixl_reg_dlist_t new_list(mem, sorted, descs.size());
                 for(long unsigned int i = 0; i<descs.size(); i++)
-                    new_list[i] = nixlBlobDesc(descs[i][0].cast<uintptr_t>(), descs[i][1].cast<size_t>(), descs[i][2].cast<uint32_t>(), descs[i][3].cast<std::string>());
+                    new_list[i] = nixlBlobDesc(descs[i][0].cast<uintptr_t>(), descs[i][1].cast<size_t>(), descs[i][2].cast<uint64_t>(), descs[i][3].cast<std::string>());
                 if (sorted) new_list.verifySorted();
                 return new_list;
             }), py::arg("type"), py::arg("descs"), py::arg("sorted")=false)
@@ -230,8 +230,8 @@ PYBIND11_MODULE(_bindings, m) {
         .def("isSorted", &nixl_reg_dlist_t::isSorted)
         .def(py::self == py::self)
         .def("__getitem__", [](nixl_reg_dlist_t &list, unsigned int i) ->
-              std::tuple<uintptr_t, size_t, uint32_t, std::string> {
-                    std::tuple<uintptr_t, size_t, uint32_t, std::string> ret;
+              std::tuple<uintptr_t, size_t, uint64_t, std::string> {
+                    std::tuple<uintptr_t, size_t, uint64_t, std::string> ret;
                     nixlBlobDesc desc = list[i];
                     std::get<0>(ret) = desc.addr;
                     std::get<1>(ret) = desc.len;
@@ -240,19 +240,19 @@ PYBIND11_MODULE(_bindings, m) {
                     return ret;
               })
         .def("__setitem__", [](nixl_reg_dlist_t &list, unsigned int i, const py::tuple &desc) {
-                list[i] = nixlBlobDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint32_t>(), desc[3].cast<std::string>());
+                list[i] = nixlBlobDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(), desc[2].cast<uint64_t>(), desc[3].cast<std::string>());
             })
         .def("addDesc", [](nixl_reg_dlist_t &list, const py::tuple &desc) {
                 list.addDesc(nixlBlobDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(),
-                                            desc[2].cast<uint32_t>(),desc[3].cast<std::string>()));
+                                            desc[2].cast<uint64_t>(),desc[3].cast<std::string>()));
             })
         .def("append", [](nixl_reg_dlist_t &list, const py::tuple &desc) {
                 list.addDesc(nixlBlobDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(),
-                                            desc[2].cast<uint32_t>(),desc[3].cast<std::string>()));
+                                            desc[2].cast<uint64_t>(),desc[3].cast<std::string>()));
             })
         .def("index", [](nixl_reg_dlist_t &list, const py::tuple &desc) {
                 int ret = list.getIndex(nixlBlobDesc(desc[0].cast<uintptr_t>(), desc[1].cast<size_t>(),
-                                                  desc[2].cast<uint32_t>(),desc[3].cast<std::string>()));
+                                                  desc[2].cast<uint64_t>(),desc[3].cast<std::string>()));
                 if(ret < 0) throw_nixl_exception((nixl_status_t) ret);
                 return ret;
             })
